==> Performance_benchmarks/Bulk_simulations/Love_dataset/3_Love_visualize.Rmd <==
---
title: "3_Love_visualize"
author: "Jeroen Gilis"
date: "05/12/2020"
output: html_document
---

**In order to run this script (3_Love_visualize.Rmd), the datasets Love_DTU_results_scaledTPM.Rdata (or, alternatively, Love_DTU_results_count.Rdata) and Love_metadata.rda are required.** These files can either be generated with the 1_Love_prepare.Rmd and 2_Love_DTU.Rmd scripts, or downloaded from Zenodo and put in the data folder of this GitHub repository.

The results of this script (FDR-TPR curves) are already available from our GitHub, under the directory ./Results/Love_benchmark/.

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load libraries

```{r,message=FALSE}
library(ggplot2)
library(iCOBRA)
```

# Load DTU performance data

```{r}
#load(file="./Data/Love_DTU_results_scaledTPM.Rdata")
load(file="./Data/Love_DTU_results_count.Rdata") # to run the visualization on the count data

loveBenchmark <- c(
    loveDtuBenchmark_satuRn,
    loveDtuBenchmark_DoubleExpSeq,
    loveDtuBenchmark_limmaDiffsplice,
    loveDtuBenchmark_edgeRdiffsplice,
    loveDtuBenchmark_DEXSeq,
    loveDtuBenchmark_DRIMSeq,
    loveDtuBenchmark_NBSplice
)

### Remove empty enteries (due to not tested - aka to many samples for reasonable runtime)
loveBenchmark  <- loveBenchmark[which(
    sapply(loveBenchmark, function(x) ! is.null(x$dtuAnalysis))
)]

rm(loveDtuBenchmark_satuRn,
    loveDtuBenchmark_DoubleExpSeq,
    loveDtuBenchmark_limmaDiffsplice,
    loveDtuBenchmark_edgeRdiffsplice,
    loveDtuBenchmark_DEXSeq,
    loveDtuBenchmark_DRIMSeq,
    loveDtuBenchmark_NBSplice)
invisible(gc())
```

# Load and wrangle truth file

```{r}
# define the truth (copy from swimming downstream paper by Love et al.,F1000Research 2018, 7:952)
load("./Data/Love_metadata.rda")
txdf <- txdf[match(rownames(tpms), txdf$TXNAME),]
txdf$dtu.genes <- iso.dtu | iso.dte & !iso.dte.only
full.dtu.genes <- unique(txdf$GENEID[txdf$dtu.genes])

txp.exprs <- rowSums(tpms) > 0
dtu.dte.genes <- unique(txdf$GENEID[iso.dte & !iso.dte.only])
txdf$full.dtu <- iso.dtu | (txdf$GENEID %in% dtu.dte.genes & txp.exprs)
dtu.txps <- txdf$TXNAME[txdf$full.dtu]
```

# Helper function 
To get pvalues for the different methods on the different datasets

```{r}
get_pvalues <- function(dataset,nrep,nmethods) {

  get_all_txs <- function(dataset,nrep){
  
    all_txs <- c()
 
    for (i in 1:nrep) {
      current_txs <- as.character(dataset[[i]]$dtuAnalysis$TXNAME)
      all_txs <- c(all_txs, current_txs)
      all_txs <- unique(all_txs)
    }
    return(all_txs)
  }

  all_txs <- get_all_txs(dataset,nrep)
  pvalues <- matrix(data=NA, nrow=length(all_txs), ncol=nmethods*nrep)
  rownames(pvalues) <- all_txs
  
  for (i in 1:(nmethods*nrep)) {
    pvalues_new <- data.frame(dataset[[i]]$dtuAnalysis$p_value)
    rownames(pvalues_new) <- dataset[[i]]$dtuAnalysis$TXNAME
    colnames(pvalues_new) <- names(dataset)[i]

    pvalues[,i] <- pvalues_new[match(rownames(pvalues),rownames(pvalues_new)),1]
  }
  
  pvalues <- apply(pvalues, 2, as.numeric)
  rownames(pvalues) <- all_txs
  pvalues <- as.data.frame(pvalues)
  pvalues <- pvalues[order(rownames(pvalues)),]
  
  colnames(pvalues) <- c("satuRn", "DoubleExpSeq", "limmaDiffsplice", "edgeRDiffsplice", "DEXSeq", "DRIMSeq", "NBSplice")

  return(pvalues)
}
```

# Visualization function

```{r}
visualize_datasets <- function(dataset,nFilters,nSampleSizes,nrep,nmethods) {
  
  nDatasets <- nFilters * nSampleSizes
  print(paste("This function will generate",nDatasets, "plots"))
  
  ##  We must select the correct dataframes from the list, based on the list names
  selection <- c("3_rep_._filterLenient","6_rep_._filterLenient","10_rep_._filterLenient","3_rep_._filterStringent","6_rep_._filterStringent","10_rep_._filterStringent")
  
  for (j in seq_len(nDatasets)) {
    print(j) # print progress

    # get current dataset
    loveBenchmark_dataset <- loveBenchmark[grepl(selection[j], names(loveBenchmark))]
    
    pvalues <- list()
    
    for (k in seq_len(nrep)) {
        
        # get current repeat of current dataset
        loveBenchmark_dataset_rep <- loveBenchmark_dataset[grep(paste0("rep_",k,"_fil"),names(loveBenchmark_dataset))]
        pvalues_current <- get_pvalues(loveBenchmark_dataset_rep,1,nmethods)

        pvalues_current[is.na(pvalues_current)] <- 1

        # to allow for calculating average results over repeats, I concatenate thte results over repeats, which requires non-identical rownames
        rownames(pvalues_current) <- paste0(rownames(pvalues_current),"_",k)
        
        pvalues[[k]] <- pvalues_current
    }

    # concatenate to get average performance over the three datasets
    pvalues_full <- rbind(pvalues[[1]],pvalues[[2]],pvalues[[3]])
    
    # generate truth file
    truth_full_new <- as.data.frame(cbind(rownames(pvalues_full),rep(0,nrow(pvalues_full))))
    colnames(truth_full_new) <- c("TXNAME","truth")
    rownames(truth_full_new) <- truth_full_new$TXNAME
    truth_full_new$truth <- as.numeric(as.character(truth_full_new$truth))
    truth_full_new[which(sub("\\_.*", "", truth_full_new$TXNAME) %in% dtu.txps),"truth"] <- 1

    # plot FDR-TPR with iCOBRA
    cobra <- COBRAData(pval = pvalues_full, truth = truth_full_new)
    cobra <- calculate_adjp(cobra)
    cobra1perf <- calculate_performance(cobra, binary_truth = "truth", cont_truth = "none", splv = "none", aspects = c("fdrtpr", "fdrtprcurve", "overlap"))
    
    cobraplot <- prepare_data_for_plot(cobra1perf, colorscheme = "Dark2", facetted = TRUE)
    
    new_col <- c(rep(c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "black", "#999999"),4),rep("white",16))
    names(new_col) <- names(cobraplot@plotcolors)
    cobraplot@plotcolors <- new_col
  
    plot_full <- plot_fdrtprcurve(cobraplot, xaxisrange = c(0, 0.35), yaxisrange = c(0,0.85))
    
    #titles <- c("3 versus 3 - edgeR filter - scaledTPM","6 versus 6 - edgeR filter - scaledTPM","10 versus 10 - edgeR filter - scaledTPM","3 versus 3 - DRIMSeq filter - scaledTPM","6 versus 6 - DRIMSeq filter - scaledTPM","10 versus 10 - DRIMSeq filter - scaledTPM")
    titles <- c("3 versus 3 - edgeR filter - count","6 versus 6 - edgeR filter - count","10 versus 10 - edgeR filter - count","3 versus 3 - DRIMSeq filter - count","6 versus 6 - DRIMSeq filter - count","10 versus 10 - DRIMSeq filter - count")
    current_title <- titles[j]

    plot_full <- plot_full + 
      ggtitle(current_title) +
      theme(strip.background = element_blank(),
            strip.text.x = element_blank(),
            plot.title = element_text(size=10),
            axis.text.x = element_text(size = 10),
            axis.text.y = element_text(size = 10),
            axis.title.x = element_text(size = 10),
            axis.title.y = element_text(size = 10))
    
    png(paste("./Results/Love_benchmark/FDRTPR_scaledTPM_", j, ".png", sep = ""),
          width     = 5.5,
          height    = 4.5,
          units     = "in",
          res       = 200,
          pointsize = 4) # start export
    print(plot_full)
    dev.off()
    
    # png(paste("./Results/Love_benchmark/FDRTPR_count_", j, ".png", sep = ""),
    #       width     = 5.5,
    #       height    = 4.5,
    #       units     = "in",
    #       res       = 200,
    #       pointsize = 4) # start export
    # print(plot_full)
    # dev.off()
  }
}
```

# Generate FDR-TPR curves

```{r}
visualize_datasets(dataset=loveBenchmark,
                   nFilters=2,
                   nSampleSizes=3,
                   nrep=3,
                   nmethods=7)
```
