==> Performance_benchmarks/Bulk_simulations/Love_dataset/4_Love_DTU_subset.Rmd <==
---
title: "4_Love_DTU_subset"
author: "Jeroen Gilis"
date: "07/12/2020"
output: html_document
---

In this additional analysis, we perform a DTU analysis on the benchmark datasets from Love et al., but only on a subset of the transcripts. This allow us to additionally benchmark BANDITS, which (in default settings) seem to scale poorly with respect to the number of transcripts in the data and could thus not be included in our other benchmarks.

**In order to run this script (4_Love_DTU_subset.Rmd), the dataset Love_benchmark_datasets_scaledTPM.Rdata (or, alternatively, Love_benchmark_datasets_count.Rdata) is required.** This file can either be generated with the 1_Love_prepare.Rmd script or downloaded from Zenodo.

Here we run the DTU analyses for 8 DTU methods on all (18) Love benchmark datasets with a reduced number of transcripts. This code runs approximately 4 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the BANDITS analysis. 

**If you do not want to run this script, its output can also be downloaded from Zenodo: Love_DTU_results_scaledTPM_subset.Rdata (or, alernatively, Love_DTU_results_count_subset.Rdata)**

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load DTU methods

Source file that allows running all seven DTU methods that were assessed in this paper: satuRn, DoubleExpSeq, limma diffsplice, edgeRdiffsplice, DEXSeq, DRIMseq and NBSplice.

```{r, message=FALSE, warning=FALSE}
### load libraries
library(edgeR)
library(limma)
library(DEXSeq)
library(DRIMSeq)
library(DoubleExpSeq)
library(NBSplice)
library(satuRn)
library(BANDITS)
library(doMC)
library(stringr)

source(file="./Performance_benchmarks/DTU_methods.R")
```

# BANDITS

```{r}
### Functions for running BANDITS
Bandits_DTU <- function(countData, tx2gene, sampleData) {
      
  ## get equivalence class counts (ECC) files
  sample_names <- c("sample_01","sample_02","sample_03","sample_04","sample_05","sample_06","sample_07","sample_08","sample_09","sample_10","sample_11","sample_12","sample_13","sample_14","sample_15","sample_16","sample_17","sample_18","sample_19","sample_20","sample_21","sample_22","sample_23","sample_24")
  
  equiv_classes_files = file.path("./Data/Love_salmon/", sample_names, "aux_info", "eq_classes.txt")

  equiv_classes_files <- equiv_classes_files[which(unlist(str_extract_all(equiv_classes_files, ("sample_[0-9]{2}"))) %in% colnames(countData))]
  
  eff_len <- readRDS("./Data/Love_eff_len.Rds") ## get median effective TX length
  
  colnames(tx2gene)[1:2] <- c("TXNAME", "GENEID")
  group <- factor(sampleData$condition)
  design <- model.matrix(~0+group)
  
  eff_len <- eff_len[which(names(eff_len) %in% rownames(countData))]
  
  tx2gene_bandits <- tx2gene[,c(2,1)]
  
  transcripts_to_keep = filter_transcripts(gene_to_transcript = tx2gene_bandits,
                                      transcript_counts = countData,
                                      min_transcript_proportion = 0,
                                      min_transcript_counts = 0,
                                      min_gene_counts = 0)
  
  input_data <- create_data(salmon_or_kallisto = "salmon",
                     gene_to_transcript = tx2gene_bandits,
                     salmon_path_to_eq_classes = equiv_classes_files,
                     eff_len = eff_len, 
                     n_cores = 1,
                     transcripts_to_keep = transcripts_to_keep)
  
  precision <- prior_precision(gene_to_transcript = tx2gene_bandits,
                        transcript_counts = countData,
                        n_cores = 1)
  
  results <- BANDITS::test_DTU(BANDITS_data = input_data,
               precision = precision$prior,
               samples_design = sampleData,
               group_col_name = "condition",
               R = 10^4, 
               burn_in = 2*10^3,
               gene_to_transcript = tx2gene_bandits,
               n_cores = 2)
  
    localRes <- results@Transcript_results[,c(2,1,3,4)]
    rownames(localRes) <- localRes$TXNAME
    colnames(localRes) <-  colnames(localRes) <- c("TXNAME", "GENEID", "p_value","FDR")
    
    return(localRes)
}
```  

# Set up parallel execution

Note that the optimal parameter setting for parallelizing the DTU analyses depends on your specific computing system.

```{r}
nrCores <- 2

if(nrCores != 1) {
    doParallel <- TRUE
    doProgress <- 'none'

    registerDoMC(cores = nrCores)
} else {
    doParallel <- FALSE
    doProgress <- 'text'
}
```

# Run the analysis for all methods on the Love benchmark datasets

Here we run the DTU analyses for 8 DTU methods on all (18) Love benchmark datasets with a reduced number of transcripts. This code runs approximately 4 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the BANDITS analysis. 

```{r}
### Load benchmark data
#load(file="./Data/Love_benchmark_datasets_scaledTPM.Rdata")
load(file="./Data/Love_benchmark_datasets_count.Rdata") # to perform the analysis on counts
loveBenchmarkData <- c(LoveBenchmarkLenient,LoveBenchmarkStringent)

## to allow for BANDITS analysis, use less transcripts
## loop over list,"filter" transcripts by random subsampling 1000 genes
set.seed(5454)
for (i in seq_along(loveBenchmarkData)) {
    loveBenchmarkData[[i]]$metaInfo <- loveBenchmarkData[[i]]$metaInfo[which(loveBenchmarkData[[i]]$metaInfo$GENEID %in% sample(loveBenchmarkData[[i]]$metaInfo$GENEID,1000)),]# sample 1000 genes
    loveBenchmarkData[[i]]$data <- loveBenchmarkData[[i]]$data[loveBenchmarkData[[i]]$metaInfo$TXNAME,]
}

### Run DTU analysis on benchmark data

print("start satuRn")

tStart <- Sys.time()
suppressWarnings(loveDtuBenchmark_satuRn <- plyr::llply(
    .data = loveBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- satuRn_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
))
difftime(Sys.time(), tStart)

print("start DoubleExpSeq")

loveDtuBenchmark_DoubleExpSeq <- plyr::llply(
    .data = loveBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DoubleExpSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design,
            quiet=FALSE
        )
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)

print("start limma diffsplice")

loveDtuBenchmark_limmaDiffsplice <- plyr::llply(
    .data = loveBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- limma_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)

print("start edgeR_diffsplice")

loveDtuBenchmark_edgeRdiffsplice <- plyr::llply(
    .data = loveBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        t0 <- Sys.time()
        localRes <- edgeR_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)

print("start DEXSeq")

tStart <- Sys.time()
loveDtuBenchmark_DEXSeq <- plyr::llply(
    .data = loveBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DEXSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DRIMSeq")

tStart <- Sys.time()
loveDtuBenchmark_DRIMSeq <- plyr::llply(
    .data = loveBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DRIMSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start NBSplice")

tStart <- Sys.time()
loveDtuBenchmark_NBSplice <- plyr::llply(
    .data = loveBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- NBSplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start BANDITS")

# BANDITS (runs 3.6 hours)
tStart <- Sys.time()
suppressWarnings(loveDtuBenchmark_BANDITS <- plyr::llply(
    .data = loveBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- Bandits_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
))
difftime(Sys.time(), tStart)

### add method name to list names for easy post-analysis
names(loveDtuBenchmark_satuRn) <- paste0('satuRn_', names(loveDtuBenchmark_satuRn))
names(loveDtuBenchmark_DoubleExpSeq)  <- paste0('DoubleExpSeq_', names(loveDtuBenchmark_DoubleExpSeq))
names(loveDtuBenchmark_edgeRdiffsplice) <- paste0('edgeR_diffsplice_', names(loveDtuBenchmark_edgeRdiffsplice))
names(loveDtuBenchmark_limmaDiffsplice) <- paste0('limma_diffsplice_', names(loveDtuBenchmark_limmaDiffsplice))
names(loveDtuBenchmark_DEXSeq)   <- paste0('DEXSeq_', names(loveDtuBenchmark_DEXSeq))
names(loveDtuBenchmark_DRIMSeq) <- paste0('DRIMSeq_', names(loveDtuBenchmark_DRIMSeq))
names(loveDtuBenchmark_NBSplice)  <- paste0('NBSplice_', names(loveDtuBenchmark_NBSplice))
names(loveDtuBenchmark_BANDITS) <- paste0('BANDITS_', names(loveDtuBenchmark_BANDITS))

### Save result
save(
    loveDtuBenchmark_satuRn,
    loveDtuBenchmark_DoubleExpSeq,
    loveDtuBenchmark_limmaDiffsplice,
    loveDtuBenchmark_edgeRdiffsplice,
    loveDtuBenchmark_DEXSeq,
    loveDtuBenchmark_DRIMSeq,
    loveDtuBenchmark_NBSplice,
    loveDtuBenchmark_BANDITS,
    file="./Data/Love_DTU_results_scaledTPM_subset.Rdata"
)

save(
    loveDtuBenchmark_satuRn,
    loveDtuBenchmark_DoubleExpSeq,
    loveDtuBenchmark_limmaDiffsplice,
    loveDtuBenchmark_edgeRdiffsplice,
    loveDtuBenchmark_DEXSeq,
    loveDtuBenchmark_DRIMSeq,
    loveDtuBenchmark_NBSplice,
    loveDtuBenchmark_BANDITS,
    file="./Data/Love_DTU_results_count_subset.Rdata"
)
```
